pyproject.toml
[tool.poetry]
name = "custom-rl-jax"
version = "0.1.0"
description = ""
authors = ["Gabriel keith <gabe.keith@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gymnasium = {extras = ["box2d", "toy-text"], version = "^0.29.1"}
jax = {extras = ["cpu"], version = "^0.4.20"}
flax = "^0.7.5"
numpy = "^1.26.1"
matplotlib = "^3.8.1"
mypy = "^1.7.0"
pandas = "^2.1.3"
pygame = "^2.5.2"
tqdm = "^4.66.1"
orbax-checkpoint = "^0.4.7"
seaborn = "^0.13.0"
optax = "^0.1.7"
optuna = "^3.5.0"
gymnax = "^0.0.6"
pyarrow = "^14.0.2"
jaxtyping = "^0.2.25"
distrax = "^0.1.5"
jax-dataclasses = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
enjoy = "custom_rl_jax.vec_policy_gradient_cs.enjoy:main"
